requirements.txt
aniso8601==3.0.2
Django==2.1.4
django-cors-headers==2.4.0
django-debug-toolbar==1.11
djangorestframework==3.9.0
drf-pdf==0.2.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
mysqlclient==1.3.14
Pillow==5.3.0
promise==2.2.1
pytz==2018.7
reportlab==3.5.12
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.2.4
